refac: move match components to a dedicated namespace
