+++ src/FootballWorldCupScoreBoard.Application/Matches/IMatchProvider.cs
@@ -1,4 +1,4 @@
-﻿namespace FootballWorldCupScoreBoard.Application;
+﻿namespace FootballWorldCupScoreBoard.Application.Matches;
 
 public interface IMatchProvider
 {
